Check notes against created addresses
