+++ src/wallet.rs
@@ -61,7 +61,6 @@ const VERSION: &[u8] = &[2, 0];
 /// funds, or staking Dusk.
 pub struct Wallet<F: SecureWalletFile + Debug> {
     wallet: Option<WalletCore<LocalStore, StateStore, Prover>>,
-    addresses: Vec<Address>,
     store: LocalStore,
     file: Option<F>,
     status: fn(status: &str),
@@ -94,19 +93,11 @@ impl<F: SecureWalletFile + Debug> Wallet<F> {
             // Generate a Store Seed type from the mnemonic Seed bytes
             let seed = store::Seed::from_reader(&mut bytes)?;
 
-            let store = LocalStore::new(seed);
-
-            // Generate the default address
-            let ssk = store
-                .retrieve_ssk(0)
-                .expect("wallet seed should be available");
-
-            let address = Address::new(0, ssk.public_spend_key());
+            let store = LocalStore::new(seed, 1);
 
             // return new wallet instance
             Ok(Wallet {
                 wallet: None,
-                addresses: vec![address],
                 store,
                 file: None,
                 status: |_| {},
@@ -172,22 +163,11 @@ impl<F: SecureWalletFile + Debug> Wallet<F> {
 
         let (seed, address_count) = result?;
 
-        let store = LocalStore::new(seed);
-
-        let addresses: Vec<_> = (0..address_count)
-            .map(|i| {
-                let ssk = store
-                    .retrieve_ssk(i as u64)
-                    .expect("wallet seed should be available");
-
-                Address::new(i, ssk.public_spend_key())
-            })
-            .collect();
+        let store = LocalStore::new(seed, address_count);
 
         // create and return
         Ok(Self {
             wallet: None,
-            addresses,
             store,
             file: Some(file),
             status: |_| {},
@@ -213,17 +193,11 @@ impl<F: SecureWalletFile + Debug> Wallet<F> {
         let seed = store::Seed::from_reader(&mut &bytes[..])
             .map_err(|_| Error::WalletFileCorrupted)?;
 
-        let store = LocalStore::new(seed);
-        let ssk = store
-            .retrieve_ssk(0)
-            .expect("wallet seed should be available");
-
-        let address = Address::new(0, ssk.public_spend_key());
+        let store = LocalStore::new(seed, 1);
 
         // return the store
         Ok(Self {
             wallet: None,
-            addresses: vec![address],
             store,
             file: Some(file),
             status: |_| {},
@@ -241,7 +215,7 @@ impl<F: SecureWalletFile + Debug> Wallet<F> {
                 // create file payload
                 let seed = self.store.get_seed()?;
                 let mut payload = seed.to_vec();
-                payload.push(self.addresses.len() as u8);
+                payload.push(self.store.addresses().len() as u8);
 
                 // encrypt the payload
                 payload = encrypt(&payload, f.pwd())?;
@@ -378,25 +352,17 @@ impl<F: SecureWalletFile + Debug> Wallet<F> {
     /// Creates a new public address.
     /// The addresses generated are deterministic across sessions.
     pub fn new_address(&mut self) -> &Address {
-        let len = self.addresses.len();
-        let ssk = self
-            .store
-            .retrieve_ssk(len as u64)
-            .expect("wallet seed should be available");
-        let addr = Address::new(len as u8, ssk.public_spend_key());
-
-        self.addresses.push(addr);
-        self.addresses.last().unwrap()
+        self.store.new_address()
     }
 
     /// Default public address for this wallet
     pub fn default_address(&self) -> &Address {
-        &self.addresses[0]
+        &self.store.addresses()[0]
     }
 
     /// Addresses that have been generated by the user
     pub fn addresses(&self) -> &Vec<Address> {
-        &self.addresses
+        self.store.addresses()
     }
 
     /// Executes a generic contract call
@@ -769,7 +735,7 @@ mod tests {
 
         assert!(format!("{}", default_addr).eq(TEST_ADDR));
         assert_ne!(&default_addr, other_addr);
-        assert_eq!(wallet.addresses.len(), 2);
+        assert_eq!(wallet.store.addresses().len(), 2);
 
         // create another wallet with different mnemonic
         let wallet: Wallet<WalletFile> = Wallet::new("demise monitor elegant cradle squeeze cheap parrot venture stereo humor scout denial action receive flat")?;
